feat: implement CLI options and file processing [deps-change]
- Add configuration schema for CLI options
- Create functions for file reading, writing, and deletion
- Implement file conversion to markdown format
- Add command executor for processing files with CLI options
- Introduce console reporter for displaying results
