+++ docs/reports/dependencies/layers.md
@@ -15,18 +15,18 @@ flowchart LR
     end
     subgraph node//modules["node_modules"]
         node//modules/valibot/dist/index.d.cts["valibot"]
+        node//modules/globby/index.d.ts["globby"]
         node//modules/cosmiconfig/dist/index.d.ts["cosmiconfig"]
         node//modules/cosmiconfig//typescript//loader/dist/types/index.d.ts["cosmiconfig-typescript-loader"]
-        node//modules/globby/index.d.ts["globby"]
         node//modules/commander/typings/index.d.ts["commander"]
     end
     src/domain-->node//modules/valibot/dist/index.d.cts
     src/application-->src/domain
+    src/infrastructure-->node//modules/globby/index.d.ts
+    src/application-->src/infrastructure
     src/infrastructure-->node//modules/cosmiconfig/dist/index.d.ts
     src/infrastructure-->node//modules/cosmiconfig//typescript//loader/dist/types/index.d.ts
     src/infrastructure-->src/domain
-    src/infrastructure-->src/application
-    src/infrastructure-->node//modules/globby/index.d.ts
     src/infrastructure-->package.json
     src/presentation-->src/application
     src/presentation-->src/domain
