+++ src/application/process-result-types.ts
